Remove icon suffix from generated components
This generator can be used for more than icons, so it would be great if the generated component name is just based on the original file name (without having `'Icon'` appended to it). If someone wants that ending, they could just rename the image file.

If you still want this functionality, another way to do it would be to add a `--suffix` CLI argument. I can make that change as well.