Input folder configuration
This extends the utility to support a custom input directory, configured by setting `-i` or `--input`. For example `svg2rn -i assets/uncompiled -o assets/icons`

As a side product of the implementation, the parsing of both input and output directories is now more robust due to parsing with `path.resolve`.